test(intent_ir): unblock baseline — drop no_std, move tests inline
Integration tests could not compile: the crate was declared `#![no_std]`
with a custom panic handler and libc allocator, and the default test
harness links std. Stable Rust refuses both of the workarounds:
  - `panic = "abort"` in profile.dev can't link with the precompiled
    libstd the harness brings in;
  - the default `panic = "unwind"` can't compile a `no_std` crate
    without `-Zbuild-std` (nightly-only).

The `no_std` posture was speculative. This crate ships only as a
`staticlib` linked into KellyFFI alongside libc++/libSystem, so std
adds no meaningful weight. Dropping it unblocks `cargo test` on stable.

The FFI safety guarantee that justified `no_std` (no panic unwinding
through the C ABI) is preserved two other ways:
  1. `panic = "abort"` remains in profile.release (what ships).
  2. T2 will wrap every `extern "C"` entry point in `catch_unwind`.

Also:
  - Moved the 9 integration tests from `tests/validator_tests.rs` into
    an inline `#[cfg(test)] mod tests` in `src/validator.rs`. Integration
    tests become a second crate and complicate the no_std/test interaction;
    unit tests live inside the lib target and Just Work.
  - Added `rlib` to crate-type so the unit-test binary can link.

cargo test (debug + release) both green, 9 passed; cargo build --release
compiles the staticlib with zero warnings. Unblocks CI baseline for
audit/rt-ffi-safety-2026q2.
