+++ engine/intent_ir/src/lib.rs
@@ -1,46 +1,16 @@
 //! Intent IR v1 - Rust Implementation
 //!
 //! Provides validation, clamping, and safe construction of IntentFrame structures.
-
-#![no_std]
-
-extern crate alloc;
-
-// Global allocator for no_std - use system allocator via libc
-use alloc::alloc::{GlobalAlloc, Layout};
-
-struct SystemAllocator;
-
-unsafe impl GlobalAlloc for SystemAllocator {
-    unsafe fn alloc(&self, layout: Layout) -> *mut u8 {
-        // Use libc malloc for actual allocation
-        extern "C" {
-            fn malloc(size: usize) -> *mut u8;
-        }
-        malloc(layout.size())
-    }
-
-    unsafe fn dealloc(&self, ptr: *mut u8, _layout: Layout) {
-        extern "C" {
-            fn free(ptr: *mut u8);
-        }
-        free(ptr);
-    }
-}
-
-#[global_allocator]
-static ALLOCATOR: SystemAllocator = SystemAllocator;
-
-// Panic handler for no_std (abort on panic)
-#[panic_handler]
-fn panic(_info: &core::panic::PanicInfo) -> ! {
-    extern "C" {
-        fn abort() -> !;
-    }
-    unsafe {
-        abort();
-    }
-}
+//!
+//! This crate was previously `#![no_std]` with a custom libc allocator and
+//! panic-abort handler. That posture was removed because:
+//!   - The shipped artifact is a `staticlib` linked into KellyFFI, which
+//!     already depends on libc++ / libSystem — std adds no meaningful weight.
+//!   - `no_std` + `panic = "unwind"` cannot compile on stable without
+//!     `-Zbuild-std`, which breaks `cargo test`.
+//! The FFI safety guarantee (no panic unwinding through the C boundary) is
+//! preserved by `panic = "abort"` in `[profile.release]` and by wrapping
+//! every `extern "C"` entry in `std::panic::catch_unwind` (see T2).
 
 pub mod types;
 pub mod validator;
